# docs/swagger.yaml
openapi: 3.0.0
info:
  title: API de Usuários
  description: Documentação da API de gerenciamento de usuários
  version: 1.0.0
servers:
  - url: http://localhost:7777
paths:
  /api/users:
    get:
      summary: Retorna todos os usuários
      description: Retorna uma lista de todos os usuários cadastrados no sistema.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
      x-log: "Operação GET definida corretamente para /api/users"
    post:
      summary: Cria um novo usuário
      description: Cria um novo usuário com base nos dados fornecidos no corpo da requisição.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: Nome de usuário do novo usuário
                email:
                  type: string
                  format: email
                  description: Endereço de e-mail do novo usuário
                password:
                  type: string
                  description: Senha do novo usuário
              required:
                - username
                - email
                - password
      responses:
        '201':
          description: Usuário criado com sucesso
        '400':
          description: Requisição inválida
      x-log: "Operação POST definida corretamente para /api/users"
  /api/users/{id}:
    get:
      summary: Retorna um usuário pelo ID
      description: Retorna os detalhes de um usuário específico com base no ID fornecido.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          description: ID do usuário a ser retornado
          schema:
            type: string
      responses:
        '200':
          description: OK
        '404':
          description: Usuário não encontrado
      x-log: "Operação GET definida corretamente para /api/users/{id}"
    put:
      summary: Atualiza os detalhes de um usuário pelo ID
      description: Atualiza os detalhes de um usuário específico com base no ID fornecido.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          description: ID do usuário a ser atualizado
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: Novo nome de usuário
                email:
                  type: string
                  format: email
                  description: Novo endereço de e-mail
                password:
                  type: string
                  description: Nova senha
              required:
                - username
                - email
                - password
      responses:
        '200':
          description: Usuário atualizado com sucesso
        '400':
          description: Requisição inválida
        '404':
          description: Usuário não encontrado
      x-log: "Operação PUT definida corretamente para /api/users/{id}"
    delete:
      summary: Exclui um usuário pelo ID
      description: Exclui um usuário específico com base no ID fornecido.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          description: ID do usuário a ser excluído
          schema:
            type: string
      responses:
        '200':
          description: Usuário excluído com sucesso
        '404':
          description: Usuário não encontrado
      x-log: "Operação DELETE definida corretamente para /api/users/{id}"
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT